refactor: move getChatContent into services
Moved entryConversation.getChatContent into service layer and
refactored code and unit tests depending on it.
